More logic Modifications
Moved some of 'logic' from the view to gamecontroller.  Not really logic, but it may be easier to move it and use it from the controller.  Made the status bar show from the start.  Updated the messages for it.  Need to enable icons or graphics on the buttons, and change the button class when left clicked and not live.  Need to create a class method that will tag surrounding visited cells when a cell is exposed.  Need to update the controller to change the button classes of all visited cells so the 'surround exposed' method is displayed when a cell is left clicked.